requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cloudinary==1.34.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.21
django-address==0.2.8
django-allauth==0.56.1
django-resized==1.0.2
django-richtextfield==1.6.1
gunicorn==21.2.0
oauthlib==3.2.2
Pillow==10.0.0
psycopg2==2.9.7
PyJWT==2.8.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.16
